+++ sdk/dotnet-client/ThriftPluginClientArguments.cs
@@ -1,3 +1,5 @@
+using Microsoft.Extensions.Logging;
+
 namespace QueryCat.Plugins.Client;
 
 public sealed class ThriftPluginClientArguments
@@ -10,5 +12,9 @@ public sealed class ThriftPluginClientArguments
 
     public string DebugServerPath { get; set; } = string.Empty;
 
-    public string DebugServerPathArgs { get; set; } = string.Empty;
+    public string DebugServerQueryText { get; set; } = string.Empty;
+
+    public string DebugServerQueryFile { get; set; } = string.Empty;
+
+    public LogLevel LogLevel { get; set; } = LogLevel.Debug;
 }